Extracted test utilities out of datastore-core into their own module.
