+++ datastore/datastore-internal-test-utils/src/nativeMain/kotlin/androidx/datastore/core/TestActuals.native.kt
@@ -30,7 +30,7 @@ actual class TestIO actual constructor(dirName: String) {
         }
     }
     private val tmpDir = FileSystem.SYSTEM_TEMPORARY_DIRECTORY / randomFileName(dirName)
-    internal actual fun <T> newFileStorage(
+    actual fun <T> newFileStorage(
         serializer: Serializer<T>,
         prefix: String
     ): StorageImpl<T> {
