+++ settings.gradle
@@ -529,6 +529,7 @@ includeProject(":datastore:datastore", [BuildType.MAIN])
 includeProject(":datastore:datastore-core", [BuildType.MAIN])
 includeProject(":datastore:datastore-core-okio", [BuildType.MAIN])
 includeProject(":datastore:datastore-compose-samples", [BuildType.COMPOSE])
+includeProject(":datastore:datastore-internal-test-utils", [BuildType.MAIN])
 includeProject(":datastore:datastore-multiprocess", [BuildType.MAIN])
 includeProject(":datastore:datastore-preferences", [BuildType.MAIN])
 includeProject(":datastore:datastore-preferences-core", [BuildType.MAIN])
